(OPTIMIZATION) reduced BatchVertex size in half

+++ nogine2-graphics/src/graphics/mod.rs
@@ -149,13 +149,13 @@ impl Graphics {
     }
 
     /// Returns the user data.
-    pub fn user_data() -> i32 {
+    pub fn user_data() -> u16 {
         let Ok(graphics) = GRAPHICS.read() else { crash!("Couldn't access Graphics singleton!") };
         return graphics.active_scope.user_data();
     }
 
     /// Sets user data.
-    pub fn set_user_data(user_data: i32) {
+    pub fn set_user_data(user_data: u16) {
         let Ok(mut graphics) = GRAPHICS.write() else { crash!("Couldn't access Graphics singleton!") };
         graphics.active_scope.set_user_data(user_data);
     }
